+++ xenon-common/src/main/java/com/vmware/xenon/common/Utils.java
@@ -314,6 +314,27 @@ public class Utils {
         return content.toString();
     }
 
+    /**
+     * Outputs a JSON representation of the given object using useHTMLFormatting to create pretty-printed,
+     * HTML-friendly JSON or compact JSON. If hideSensitiveFields is set the JSON will not include fields
+     * with the annotation {@link com.vmware.xenon.common.ServiceDocumentDescription.PropertyUsageOption.SENSITIVE}.
+     * If hideSensitiveFields is set and the Object is a string with JSON, sensitive fields cannot be discovered will
+     * throw an Exception.
+     */
+    public static String toJson(boolean hideSensitiveFields, boolean useHtmlFormatting, Object body)
+            throws IllegalArgumentException {
+        if (body instanceof String) {
+            if (hideSensitiveFields) {
+                throw new IllegalArgumentException("Body is already a string, sensitive fields cannot be discovered");
+            }
+            return (String) body;
+        }
+        StringBuilder content = getBuilder();
+        JsonMapper mapper = getJsonMapperFor(body);
+        mapper.toJson(hideSensitiveFields, useHtmlFormatting, body, content);
+        return content.toString();
+    }
+
     public static <T> T fromJson(String json, Class<T> clazz) {
         return getJsonMapperFor(clazz).fromJson(json, clazz);
     }